[#601] BaseBatchProcessor に処理結果集計のための返り値サポートを追加 Fixes #88
