+++ src/portrait_quality_batch_processor.py
@@ -31,8 +31,13 @@ class PortraitQualityBatchProcessor(BaseBatchProcessor):
         """
         super().__init__(config_path=config_path, max_workers=max_workers)
 
+        # Base が持つ logger/config_manager を上書きしない方針に寄せる。
+        # ただし互換のため、外部 logger が渡された場合はそれを採用。
         self.config = self.config_manager.get_config()
-        self.logger = logger or self.config_manager.get_logger("PortraitQualityBatchProcessor")
+        if logger is not None:
+            self.logger = logger
+        else:
+            self.logger = self.config_manager.get_logger("PortraitQualityBatchProcessor")
         self.memory_monitor = MemoryMonitor(self.logger)
         self.date = date
         self.processed_images = set()
@@ -50,6 +55,11 @@ class PortraitQualityBatchProcessor(BaseBatchProcessor):
         self.memory_threshold_exceeded = False
 
         self.batch_size = batch_size or self.config.get("batch_size", 10)
+        self.memory_threshold = self.config_manager.get_memory_threshold(default=90)
+
+        # run bookkeeping（運用ログ用）
+        self._start_processed_count: int = 0
+        self._total_images_to_process: int = 0
 
     def on_config_change(self, new_config: dict) -> None:
         """
@@ -72,20 +82,14 @@ class PortraitQualityBatchProcessor(BaseBatchProcessor):
         self._set_directories_and_files()
         self._load_processed_images()
 
-        super().setup()
+        # ★開始時点の処理済み数（累計）を保存して差分で数える
+        self._start_processed_count = len(self.processed_images)
 
-        if self.processed_images:
-            self.logger.info(
-                f"Found {len(self.processed_images)} previously processed images. Resuming from there."
-            )
-        else:
-            self.logger.info("No previously processed images found. Starting fresh.")
+        # Base契約: setup() -> self.data = self.get_data() -> after_data_loaded(self.data)
+        super().setup()
 
         self.memory_threshold_exceeded = False
         self.completed_all_batches = False
-
-        self.memory_threshold = self.config_manager.get_memory_threshold(default=90)
-        self.logger.info(f"Memory usage threshold set to {self.memory_threshold}% from config.")
 
     def _set_directories_and_files(self) -> None:
         """
@@ -178,44 +182,23 @@ class PortraitQualityBatchProcessor(BaseBatchProcessor):
             self.logger.error(f"CSV parsing error in {self.image_csv_file}: {e}")
         return image_data
 
-    def execute(self, target_dir: Optional[Path] = None) -> None:
+    def after_data_loaded(self, data: List[Dict]) -> None:
         """
-        バッチ処理の実行エントリポイント。
-        メモリ閾値を超えた場合は警告ログを残してバッチ単位で中断する。
-        中断時には処理済み画像数もログに出力。
+        データロード後に一度だけ行う副作用をここへ集約（Base.setup() から呼ばれる）。
+        - 処理済みの存在ログ
+        - メモリ閾値ログ
+        - 対象件数ログ（未処理のみ data に入る想定）
         """
-        try:
-            self.setup()
+        self._total_images_to_process = len(data)
+        if self.processed_images:
+            self.logger.info(
+                f"Found {len(self.processed_images)} previously processed images. Resuming from there."
+            )
+        else:
+            self.logger.info("No previously processed images found. Starting fresh.")
 
-            # ★開始時点の処理済み数（累計）を保存して差分で数える
-            start_processed_count = len(self.processed_images)
-            
-            data = self.get_data(target_dir=target_dir)
-
-            total_images = len(data)
-            self.logger.info(f"Total images to process: {total_images}")
-
-            self.process(data)
-
-            processed_this_run = len(self.processed_images) - start_processed_count
-            remaining_count = total_images - processed_this_run
-
-            if getattr(self, "memory_threshold_exceeded", False):
-                self.logger.warning(
-                    f"Batch processing stopped due to memory threshold. "
-                    f"Processed {processed_this_run} images, {remaining_count} remaining. "
-                    "You can re-run to continue remaining images."
-                )
-            elif getattr(self, "completed_all_batches", False):
-                self.logger.info(
-                    f"All batches processed successfully. Total processed images: {processed_this_run}"
-                )
-
-        except Exception as e:
-            self.handle_error(f"Unexpected error during execution: {e}", raise_exception=False)
-
-        finally:
-            self.cleanup()
+        self.logger.info(f"Memory usage threshold set to {self.memory_threshold}% from config.")
+        self.logger.info(f"Total images to process: {self._total_images_to_process}")
 
     def _process_batch(self, batch: List[Dict[str, str]]) -> List[Dict[str, Any]]:
         """
@@ -419,23 +402,38 @@ class PortraitQualityBatchProcessor(BaseBatchProcessor):
         self.logger.info(
             "Cleaning up PortraitQualityBatchProcessor-specific resources."
         )
+
+        # ===== 運用ログ（旧 execute 相当）=====
+        processed_this_run = len(self.processed_images) - int(self._start_processed_count or 0)
+        remaining_count = int(self._total_images_to_process or 0) - processed_this_run
+        if remaining_count < 0:
+            remaining_count = 0
+
+        if getattr(self, "memory_threshold_exceeded", False):
+            self.logger.warning(
+                f"Batch processing stopped due to memory threshold. "
+                f"Processed {processed_this_run} images, {remaining_count} remaining. "
+                "You can re-run to continue remaining images."
+            )
+        elif getattr(self, "completed_all_batches", False):
+            self.logger.info(
+                f"All batches processed successfully. Total processed images: {processed_this_run}"
+            )
+
         processed_file = getattr(self, "processed_images_file", None)
         if getattr(self, "completed_all_batches", False) and processed_file and os.path.exists(processed_file):
             self.logger.info(f"Removing processed images file: {processed_file}")
             os.remove(processed_file)
 
-    def get_data(self, target_dir: Optional[Path] = None) -> List[Dict[str, str]]:
+    def load_data(self) -> List[Dict[str, str]]:
         """
-        処理対象の画像データのうち、未処理のものだけを返す。
-
+        BaseBatchProcessor の新契約:
+        - load_data(): 純I/O（副作用なし）
+        - キャッシュは Base が握る（get_data() は Base 側）
+ 
         Returns:
-            List[Dict[str, str]]: 未処理画像のメタデータ一覧
+            List[Dict[str, str]]: 画像メタデータ一覧（未処理のものだけ）
         """
-        if target_dir:
-            self.logger.warning(
-                f"[get_data] target_dir={target_dir} は未対応のため無視されます。"
-            )
-
         raw_data = self.load_image_data()
         return [d for d in raw_data if d["file_name"] not in self.processed_images]
